Add script to run ESMFold on yeast_ppi dataset
This commit adds a new script, esmfold.sh, which allows running ESMFold on the yeast_ppi dataset. The script takes a fasta file as input and generates PDB files using ESMFold. The resulting PDB files are saved in the esmfold_pdb directory. This script is useful for analyzing protein-protein interactions in yeast.